Add files via upload
New Terms approved Nov 28. Added a more fulsome reservation of rights section (including the right to reject any or all proposals and to cancel at any time without award, or re-issue (section 20);
·         Added the right to avoid contracting with someone who’s identity you cannot verify (section 22);
·         Added that the Issuer will not be liable for errors or omissions in the Opportunity (section 31);
 
In the Contract Terms, we've also altered the definition of Submission Period, to reflect either a mandatory completion date, or the proposed timing of the successful Developer (if that criteria is being evaluated).